requirements.txt
streamlit
google-genai
web3
pywebview
python-dotenv